requirements.txt
aiogram
requests
python-decouple
playwright